+++ packages/core/src/input-binding/boolean/plugin.ts
@@ -8,11 +8,11 @@ import {Constraint} from '../../common/constraint/constraint';
 import {ListConstraint} from '../../common/constraint/list';
 import {ListController} from '../../common/controller/list';
 import {boolFromUnknown} from '../../common/converter/boolean';
+import {createListConstraint, parseListOptions} from '../../common/list-util';
 import {parseRecord} from '../../common/micro-parsers';
 import {ValueMap} from '../../common/model/value-map';
 import {BaseInputParams, ListParamsOptions} from '../../common/params';
 import {writePrimitive} from '../../common/primitive';
-import {createListConstraint, parseListOptions} from '../../common/util';
 import {VERSION} from '../../version';
 import {InputBindingPlugin} from '../plugin';
 import {CheckboxController} from './controller/checkbox';
